project/requirements.txt
fastapi
uvicorn[standard]
python-multipart
graphrag
lancedb
pyarrow
openai
python-dotenv
